# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aldis API",
        "description": "API documentation for Aldis application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/user-locations": {
            "get": {
                "tags": [
                    "User Locations"
                ],
                "summary": "Get all user locations",
                "description": "Retrieve a list of all user locations.",
                "operationId": "1d57958d8500e47816675d278941af05",
                "responses": {
                    "200": {
                        "description": "List of user locations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User Locations"
                ],
                "summary": "Create a new user location",
                "description": "Add a new user location to the system.",
                "operationId": "b2cf17026e3c38e76295fa7f9448e588",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, City, Country"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User location created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User location created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/user-locations/{id}": {
            "get": {
                "tags": [
                    "User Locations"
                ],
                "summary": "Get a user location by ID",
                "description": "Retrieve details of a specific user location by its ID.",
                "operationId": "363441cbd1e70b58a753ed8a64a7e6e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User location ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User location details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User location not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "User Locations"
                ],
                "summary": "Update a user location",
                "description": "Update details of an existing user location.",
                "operationId": "8d9bf752d7dffac774234020036430de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User location ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "Updated Address"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User location updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User location updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User location not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User Locations"
                ],
                "summary": "Delete a user location",
                "description": "Remove a user location from the system.",
                "operationId": "1c642f163f90a04be137b2b7d9bad925",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User location ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User location deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User location deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User location not found"
                    }
                }
            }
        }
    },
...
}